+++ src/components/animation/animations/index.ts
@@ -1,12 +1,1 @@
-import animatecss from './animatecss';
-import animista from './animista';
-
-const sortObject = (o: any) =>
-  Object.keys(o)
-    .sort()
-    .reduce((r, k) => ((r[k] = o[k]), r), {});
-
-export default {
-  none: [],
-  ...sortObject({ ...animatecss, ...animista })
-};
+export * as animations from '@shoelace-style/animations';