+++ docs/getting-started/changelog.md
@@ -6,6 +6,13 @@ Components with the <sl-badge type="warning" pill>Experimental</sl-badge> badge 
 
 _During the beta period, these restrictions may be relaxed in the event of a mission-critical bug._ 🐛
 
+## Next
+
+- 🚨 BREAKING CHANGE: Fixed animations bloat
+  - Removed ~400 baked-in Animista animations because they were causing ~200KB of bloat (they can still be used with custom keyframes)
+  - Reworked animations into a separate module ([`@shoelace-style/animations`](https://github.com/shoelace-style/animations)) so it's more maintainable and animations are sync with the latest version of animate.css
+  - Animation and easing names are now camelcase (e.g. `easeInOut` instead of `ease-in-out`)
+
 ## 2.0.0-beta.25
 
 - 🚨 BREAKING CHANGE: Reworked color tokens